Updated android file
Don’t know why Git decided to hold on to an old version of an untracked
file…
